refactor: Update dependencies and add react-quill for rich text editing
