+++ package.json
@@ -23,7 +23,7 @@
     "postcss": "^8.4.38",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-pdf": "^9.1.0",
+    "react-quill": "^2.0.0",
     "react-router-dom": "^6.23.1",
     "react-scripts": "5.0.1",
     "react-time-ago": "^7.3.3",
